+++ agrest-engine/src/main/java/io/agrest/encoder/DataResponseEncoder.java
@@ -9,12 +9,17 @@ import java.io.IOException;
  */
 public class DataResponseEncoder implements Encoder {
 
-    private CollectionEncoder resultEncoder;
-    private Encoder totalEncoder;
-    private String resultProperty;
-    private String totalProperty;
+    private final CollectionEncoder resultEncoder;
+    private final Encoder totalEncoder;
+    private final String resultProperty;
+    private final String totalProperty;
 
-    public DataResponseEncoder(String resultProperty, CollectionEncoder resultEncoder, String totalProperty, Encoder totalEncoder) {
+    public DataResponseEncoder(
+            String resultProperty,
+            CollectionEncoder resultEncoder,
+            String totalProperty,
+            Encoder totalEncoder) {
+
         this.totalProperty = totalProperty;
         this.resultEncoder = resultEncoder;
         this.totalEncoder = totalEncoder;
@@ -39,10 +44,5 @@ public class DataResponseEncoder implements Encoder {
     protected void encodeObjectBody(Object object, JsonGenerator out) throws IOException {
         int count = resultEncoder.encodeAndGetTotal(resultProperty, object, out);
         totalEncoder.encode(totalProperty, count, out);
-    }
-
-    @Override
-    public int visitEntities(Object object, EncoderVisitor visitor) {
-        return resultEncoder.visitEntities(object, visitor);
     }
 }