+++ agrest-engine/src/main/java/io/agrest/encoder/EncodableProperty.java
@@ -10,7 +10,7 @@ import io.agrest.property.PropertyReader;
 public class EncodableProperty {
 
     private Encoder encoder;
-    private PropertyReader reader;
+    private final PropertyReader reader;
 
     public static EncodableProperty property(PropertyReader reader) {
         return new EncodableProperty(reader, GenericEncoder.encoder());
@@ -33,10 +33,5 @@ public class EncodableProperty {
 
     public PropertyReader getReader() {
         return reader;
-    }
-
-    public int visit(Object root, EncoderVisitor visitor) {
-        Object value = root == null ? null : getReader().value(root);
-        return encoder.visitEntities(value, visitor);
     }
 }